Rename subcomponents
